% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_data.R
\name{gc.data}
\alias{gc.data}
\title{Compile data set for growth curve fitting}
\usage{
gc.data(data, var.OD, min = NULL, max = NULL, interval = NULL)
}
\arguments{
\item{data}{Raw data from microplate readers or data in similar format}

\item{var.OD}{Variable name of OD measure(s). If multiple columns exist, use c() for input.}

\item{min}{The starting point of bacterial growth measurement}

\item{max}{The end-time of bacterial growth measurement}

\item{interval}{Time interval of data collections. Make sure the units are the same in all three time inputs.}
}
\value{
A data set with 4 time variables: t, t^2, t^3, and t^4.
}
\description{
This function convert raw data from microplate readers to a long format time-structured data set that can be used to fit polynomial growth curve.
}
\details{
This function only takes time-structured data, i.e. OD needs to be measured at the same time interval. It will provide a long format data set with 4 time variables: t, t^2, t^3, and t^4. Missing data is allowed.
}
\examples{
df.test <- gc.data(df, "C1", min = 20/60, max = 1100/60, interval = 20/60)
df.test <- gc.data(df, c("C1", "C2", "C3"), min = 20/60, max = 1100/60, interval = 20/60)



}
