% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_parameters.R
\name{gc.parameters}
\alias{gc.parameters}
\title{Calculate growth curve parameters}
\usage{
gc.parameters(model.curve, df.input, min.time, max.time)
}
\arguments{
\item{model.curve}{A polynomial regression model}

\item{df.input}{A data set with observed OD values}

\item{min.time}{Starting time}

\item{max.time}{Ending time}
}
\value{
A data set with growth curve parameters
}
\description{
Calculate growth curve parameters
}
\examples{
df.test.result <- gc.parameters(lm1, df.test, 20/60, 1100/60)

}
