% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_plot.R
\name{gc.plot}
\alias{gc.plot}
\title{Plot polynomial growth curve}
\usage{
gc.plot(gcdata, model.curve, color = "blue", title = NULL)
}
\arguments{
\item{gcdata}{A data set from gc.data()}

\item{model.curve}{A polynomial regression model}

\item{color}{Color of fitted polynomial growth curve. The default color is blue.}

\item{title}{Title of the plot. If not specified, there will be no title.}
}
\value{
A plot
}
\description{
This function plot the observed bacterial growth data with a line of fitted polynomial growth curve.
}
\examples{
gc.plot(df.test, lm1)
gc.plot(df.test, lm1, color = "red", title = "My curve")



}
