% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_poly.R
\name{gc.poly}
\alias{gc.poly}
\title{Fit polynomial growth curve}
\usage{
gc.poly(gcdata)
}
\arguments{
\item{gcdata}{A data set from gc.data()}
}
\value{
A model object
}
\description{
Fit polynomial growth curve
}
\details{
This function fits polynomial growth curve and outputs a model object that can be used for
parameter estimation and plot.
}
\examples{
lm1 <- gc.poly(df.test)



}
