% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipg_multisample.R
\name{ipg_multisample}
\alias{ipg_multisample}
\title{Growth parameter estimates for multiple samples}
\usage{
ipg_multisample(df.input, factors, time.name, y.name, max.y.select)
}
\arguments{
\item{df.input}{Input data containing the time and dependent variable from a single sample. Data needs to be in long format.
Remove rows with missing in the dependent variable.}

\item{factors}{Sample identifier(s). When exists more than one identifier, put a vector of variable names of the identifiers.}

\item{time.name}{Name of the time variable}

\item{y.name}{Name of the dependent variable}

\item{max.y.select}{A threshold for maximum detection of dependent variable in exponential phase. Default is 0.2\%.
This number will be multiplied by the range of the dependent variable to establish a threshold for maximum detection.}
}
\value{
A list that contains a table of estimates, polynomial models, a table of beta coefficients, and a table of fitted values, all by sample identifier(s).
Growth parameters include peak growth rate, peak growth time, lag time, max y (fit), max y time (fit), and doubling time at the peak growth.
}
\description{
This function estimates growth parameters for a dataset with multiple samples. Technical replicates are allowed in the dataset.
}
\details{
The function uses the same approach to estimate growth parameters as in \code{ipg_singlesample()}.
}
\examples{
library(dplyr)
data <- growthrates::bactgrowth
out.multisample <- ipg_multisample(data, c("strain", "conc"), "time", "value", 0.2/100)
}
