% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipolygrowth-package.R
\docType{package}
\name{ipolygrowth-package}
\alias{ipolygrowth}
\alias{ipolygrowth-package}
\title{ipolygrowth: Bacterial Growth Curve Parameters from Polynomial Model}
\description{
Taking data from microplate reader and estimating bacterial grwoth parameters using a 4th-degree polynomial regression model.
}
\details{
This package estimates bacterial growth curve parameters using the fourth-degree polynomial model.
Data from microplate readers can be used directly.
The package is capable of taking either a single sample or multiple samples, with or without replicates.
Parameter estimates include peak growth rate, peak growth time, lag time, max y (fit), max y time (fit), and doubling time at the peak growth.
Although the package is designed for bacterial data, the use of this package may be expanded to other growth curve data.
}
\references{
Cross JH, Bradbury RS, Fulford AJ, Jallow AT, Wegmüller R, Prentice AM, Cerami C. Oral iron acutely elevates bacterial growth in human serum. Sci Rep. 2015 Nov 23;5:16670. doi: 10.1038/srep16670. PMID: 26593732; PMCID: PMC4655407.
ipolygrowth publication
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kivanvan/ipolygrowth}
  \item Report bugs at \url{https://github.com/kivanvan/ipolygrowth/issues}
}

}
\author{
\strong{Maintainer}: Jifan Wang \email{jifan.wang@tufts.edu} (\href{https://orcid.org/0000-0001-9720-8913}{ORCID})

Authors:
\itemize{
  \item Kathryn Barger
}

}
\keyword{internal}
