% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipg_singlesample.R
\name{ipg_singlesample}
\alias{ipg_singlesample}
\title{Growth parameter estimates for single sample}
\usage{
ipg_singlesample(df.input, time.name, y.name, max.y.select = 0.2/100)
}
\arguments{
\item{df.input}{Input data containing the time and dependent variable from a single sample. Data needs to be in long format.}

\item{time.name}{Name of the time variable}

\item{y.name}{Name of the dependent variable}

\item{max.y.select}{A threshold for maximum detection of dependent variable in exponential phase. Default is 0.2\%.
This number will be multiplied by the range of the dependent variable to establish a threshold for maximum detection.}
}
\value{
A list that contains a table of estimates, the polynomial model, a table of beta coefficients, and a table of fitted values.
Growth parameters include peak growth rate, peak growth time, lag time, max y (fit), max y time (fit), and doubling time at the peak growth.
}
\description{
This function estimates growth parameters for a single sample. Technical replicates are allowed in the dataset.
}
\details{
This function fits a 4th-degree polynomial model on input data. Peak growth time is first determined by the second derivative of model.
Peak growth rate is then calculated as the first derivative at peak growth time.
Doubling time at the peak growth is calculated using the equation: \eqn{\frac{ln(2)}{peak\ growth\ rate}}.
The lag time is estimated as the time when the line of intercept from the polynomial model intersects with the linear line of peak growth at peak growth time.
Difference in fitted value of y between two successive time intervals is compared to the selected threshold (the input percentage * range of y) sequentially to estimate the maximum of y and its time.
}
\examples{
data <- growthrates::bactgrowth
df.singlesample <- data \%>\% dplyr::filter(strain == "D", conc == 0)
out.singlesample <- ipg_singlesample(df.singlesample, "time", "value")


}
