% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipolygrowth-package.R
\docType{package}
\name{ipolygrowth-package}
\alias{ipolygrowth}
\alias{ipolygrowth-package}
\title{ipolygrowth: Individual growth curve parameter calculation using polynomial functions}
\description{
Calculation of bacterial growth curve parameters using fourth degree polynomial functions.
}
\details{
This package calculates growth curve parameters using fourth degree polynomial functions.
The package was developed for bacterial growth experiments.
Thus, data from microplate readers serve as the input data format for R functions in this package.
Polynomial growth curves are estimated from time series data from a single biological sample or multiple samples.
Technical replicates within biological samples are allowed.
Output includes six calculated growth curve parameters: peak growth rate, peak growth time, doubling time at peak growth, lag time, max y, and max y time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kivanvan/ipolygrowth}
  \item Report bugs at \url{https://github.com/kivanvan/ipolygrowth/issues}
}

}
\author{
\strong{Maintainer}: Jifan Wang \email{jifan.wang@tufts.edu} (\href{https://orcid.org/0000-0001-9720-8913}{ORCID})

Authors:
\itemize{
  \item Kathryn Barger
}

}
\keyword{internal}
